+++ BxInputController/Sources/Rows/Text/BxInputSelectorTextRow.swift
@@ -38,10 +38,10 @@ open class BxInputSelectorTextRow : BxInputTextRow, BxInputSelectorRow
     }
     
     override public init(title: String? = nil, subtitle: String? = nil,
-                         placeholder: String? = nil, value: String? = nil)
+                         placeholder: String? = nil, maxCount: Int? = nil, value: String? = nil)
     {
         child = BxInputChildSelectorTextRow()
-        super.init(title: title, subtitle: subtitle, placeholder: placeholder, value: value)
+        super.init(title: title, subtitle: subtitle, placeholder: placeholder, maxCount: maxCount, value: value)
         child.parent = self
     }